fix(macos): cast statvfs f_blocks/f_bfree to u64 for saturating_mul
- On macOS statvfs returns u32 for f_blocks and f_bfree, causing type mismatch
  with saturating_mul which expects u64 on both sides
- Linux uses unsigned long (u64 on 64-bit) so no change needed there

+++ src-tauri/src/drive/macos.rs
@@ -126,8 +126,8 @@ fn get_capacity(path: &Path) -> (u64, u64) {
     }
 
     let frag_size = stat.f_frsize.max(1) as u64;
-    let total = frag_size.saturating_mul(stat.f_blocks);
-    let free = frag_size.saturating_mul(stat.f_bfree);
+    let total = frag_size.saturating_mul(u64::from(stat.f_blocks));
+    let free = frag_size.saturating_mul(u64::from(stat.f_bfree));
     (total, free)
 }
